added showing event resources
Added recycler view for event resources which are shown in the EventDetailsFragment.

+++ app/src/main/java/com/nikitasutulov/macsro/data/dto/operations/resourcesevent/CreateResourcesEventDto.kt
@@ -3,6 +3,6 @@ package com.nikitasutulov.macsro.data.dto.operations.resourcesevent
 data class CreateResourcesEventDto(
     val resourceGID: String,
     val eventGID: String,
-    val requiredQuantity: Int,
-    val availableQuantity: Int
+    val requiredQuantity: Double,
+    val availableQuantity: Double
 )
