+++ app/src/main/java/com/nikitasutulov/macsro/data/dto/operations/resourcesevent/ResourcesEventDto.kt
@@ -4,6 +4,6 @@ data class ResourcesEventDto(
     val gid: String,
     val resourceGID: String,
     val eventGID: String,
-    val requiredQuantity: Int,
-    val availableQuantity: Int
+    val requiredQuantity: Double,
+    val availableQuantity: Double
 )
